requirements.txt
absl-py==2.1.0
altair==5.5.0
anyio==4.9.0
appdirs==1.4.4
argon2-cffi==21.1.0
asgiref==3.8.1
astunparse==1.6.3
attrs==21.2.0
Babel==2.8.0
backcall==0.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
bleach==4.1.0
blinker==1.9.0
boto3==1.34.153
botocore==1.34.153
branca==0.8.1
Brotli==1.0.9
CacheControl==0.14.2
cachetools==5.5.2
certifi==2024.2.2
chardet==4.0.0
charset-normalizer==3.3.2
ciso8601==2.3.2
cliapp==1.20180812.1
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cmdtest==0.32+git
colorama==0.4.6
command-not-found==0.3
cryptography==3.4.8
cycler==0.11.0
dbus-python==1.2.18
debugpy==1.8.13
decorator==4.4.2
defusedxml==0.7.1
distro==1.7.0
distro-info==1.1+ubuntu0.1
Django==5.0.7
djangorestframework==3.15.2
dnspython==2.4.2
entrypoints==0.4
exceptiongroup==1.2.2
executing==2.0.0
fastapi==0.97.0
fiona==1.10.1
firebase==4.0.1
firebase-admin==6.7.0
Flask==2.2.5
Flask-Cors==4.0.1
Flask-PyMongo==2.3.0
flatbuffers==24.3.25
folium==0.14.0
fonttools==4.29.1
fs==2.4.12
gast==0.5.4
gcloud==0.18.3
GDAL==3.4.1
geopandas==1.0.1
-e git+https://github.com/chakri0176/git-storyteller@a5f5dca2689133c7b07fa8b9f3a382da46c68238#egg=git_storyteller
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.2
google-api-python-client==2.164.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.3
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.6.0
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.1
googlemaps==4.10.0
grpcio==1.71.0
grpcio-status==1.62.3
h11==0.16.0
h5py==3.11.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.20.2
httptools==0.6.4
httpx==0.26.0
idna==3.7
importlib-metadata==4.6.4
iniconfig==2.1.0
ipykernel==6.7.0
ipython==7.31.1
ipython_genutils==0.2.0
ipywidgets==6.0.0
itsdangerous==2.2.0
jedi==0.18.0
jeepney==0.7.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.4.2
jsonify==0.5
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
keras==2.15.0
keyring==23.5.0
kiwisolver==1.3.2
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==18.1.1
lxml==4.8.0
lz4==3.1.3+dfsg
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.5.1
matplotlib-inline==0.1.3
mdurl==0.1.2
ml-dtypes==0.2.0
mongoengine==0.28.2
more-itertools==8.10.0
mpmath==0.0.0
msgpack==1.1.0
namex==0.0.8
narwhals==1.30.0
nbclient==0.5.6
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.4
netifaces==0.11.0
nltk==3.9.1
notebook==6.4.8
npm==0.1.1
numpy==1.24.3
oauth2client==4.1.3
oauthlib==3.2.0
olefile==0.46
opencv-python==4.9.0.80
opt-einsum==3.3.0
optional-django==0.1.0
optree==0.11.0
packaging==21.3
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pluggy==1.5.0
ply==3.11
prometheus-client==0.9.0
prompt-toolkit==3.0.28
proto-plus==1.26.1
protobuf==4.25.6
ptyprocess==0.7.0
py==1.10.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==1.10.21
pydeck==0.9.1
pyee==8.2.2
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.10.1
pymongo==4.8.0
pyogrio==0.10.0
pyparsing==2.4.7
pyppeteer==1.0.2
pyproj==3.7.1
pyrsistent==0.18.1
pytesseract==0.3.10
pytest==8.3.5
python-apt==2.4.0+ubuntu2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.6
pythran==0.10.0
pytz==2022.1
PyYAML==5.4.1
pyzmq==22.3.0
regex==2024.11.6
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
s3transfer==0.10.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
SecretStorage==3.3.1
Send2Trash==1.8.1b0
shapely==2.0.1
shellingham==1.5.4
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.3.1
sqlparse==0.5.1
starlette==0.27.0
streamlit==1.43.1
sympy==1.9
systemd-python==234
tenacity==9.0.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.0
termcolor==2.4.0
terminado==0.13.1
testpath==0.5.0
threadpoolctl==3.5.0
toml==0.10.2
tomli==2.2.1
tornado==6.1
tqdm==4.66.2
traitlets==5.1.1
tsx==0.2.0
ttystatus==0.38
typer==0.15.3
typing_extensions==4.11.0
tzdata==2024.1
ubuntu-advantage-tools==8001
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
uritemplate==4.1.1
urllib3==1.26.18
uvicorn==0.22.0
uvloop==0.21.0
varname==0.12.0
wadllib==1.3.6
waitress==3.0.2
watchdog==6.0.0
watchfiles==1.0.5
wcwidth==0.2.5
webencodings==0.5.1
websockets==10.4
Werkzeug==2.2.3
widgetsnbextension==2.0.0
wrapt==1.14.1
zipp==1.0.0
